- Jorrit fixed the memory tracker version of the block allocator 	  so that it is 64-bit clean. 	- Jorrit added a work around for the placement new as used in 	  the GL renderer so that it compiles with the memory tracker 	  enabled.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@23337 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csutil/blockallocator.h
@@ -108,8 +108,8 @@ public:
   {
     char buf[255];
     sprintf (buf, "csBlockAllocator<%s>", typeid (T).name());
-    int32* ptr = (int32*)malloc (blocksize + sizeof (int32)*2);
-    *ptr++ = (int32)mtiRegisterAlloc (blocksize, buf);
+    uintptr_t* ptr = (uintptr_t*)malloc (blocksize + sizeof (uintptr_t)*2);
+    *ptr++ = (uintptr_t)mtiRegisterAlloc (blocksize, buf);
     *ptr++ = blocksize;
     return (uint8*)ptr;
   }
@@ -120,7 +120,7 @@ public:
    */
   static inline void FreeBlock(uint8* p)
   {
-    int32* ptr = ((int32*)p)-2;
+    uintptr_t* ptr = ((uintptr_t*)p)-2;
     mtiRegisterFree ((csMemTrackerInfo*)*ptr, (size_t)ptr[1]);
     free (ptr);
   }
